+++ e2e/E2E_USAGE.md
@@ -5,7 +5,7 @@ This guide explains how to run the **storage-e2e** smoke tests for `sds-node-con
 ## What the test does
 
 - Uses the [storage-e2e](https://github.com/deckhouse/storage-e2e) framework.
-- Entry point: `TestE2E` in `e2e/tests/e2e_suite_test.go` (single Ginkgo run: Common Scheduler scenarios first, then Sds Node Configurator — see file order / `Describe` order).
+- Entry point: `TestSdsNodeConfigurator` in `e2e/tests/sds_node_configurator_suite_test.go` (single Ginkgo run; nested cluster in `BeforeSuite`; Common Scheduler + Sds specs in package).
 - Covers BlockDevice discovery and LVMVolumeGroup flows on a test cluster (existing or created via framework).
 - Test cluster is reached via SSH (base cluster host or jump host) and kubeconfig (often through an SSH tunnel to the API).
 
@@ -101,7 +101,7 @@ ginkgo -v --progress ./tests/
 Or run specific test:
 
 ```bash
-# Ginkgo focus on a spec name; CI runs: go test ./tests/ -run '^TestE2E$'
+# Ginkgo focus on a spec name; CI runs: go test ./tests/ -run '^TestSdsNodeConfigurator$'
 ginkgo -v --progress --focus="Should schedule Pod with local PVC" ./tests/
 ```
 
@@ -173,4 +173,4 @@ storage-e2e checks the Deckhouse `Module/virtualization` once with a short timeo
 ## See also
 
 - [README.md](README.md) — test scenarios, debugging, troubleshooting.
-- Local runs: `ginkgo -v --progress ./tests/` or `go test -v -count=1 -timeout 60m ./tests/ -run '^TestE2E$'` (same as CI).
+- Local runs: `ginkgo -v --progress ./tests/` or `go test -v -count=1 -timeout 60m ./tests/ -run '^TestSdsNodeConfigurator$'` (same as CI).
