+++ e2e/tests/sds_node_configurator_suite_test.go
@@ -5,7 +5,7 @@ Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+	http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
@@ -29,6 +29,8 @@ import (
 var _ = BeforeSuite(func() {
 	err := setup.Init()
 	Expect(err).NotTo(HaveOccurred(), "Failed to initialize storage-e2e setup")
+	// Before any spec: Ginkgo may shuffle root Ordered Describes; nested cluster must exist first.
+	e2eEnsureSharedNestedTestCluster()
 })
 
 var _ = AfterSuite(func() {
@@ -38,17 +40,13 @@ var _ = AfterSuite(func() {
 	}
 })
 
-// TestE2E is the single entry point for all Ginkgo specs in this package.
-// Spec order follows file registration order: common_scheduler_test.go (Common Scheduler Extender)
-// runs before sds_node_configurator_test.go (Sds Node Configurator).
-func TestE2E(t *testing.T) {
+func TestSdsNodeConfigurator(t *testing.T) {
 	RegisterFailHandler(Fail)
 	suiteConfig, reporterConfig := GinkgoConfiguration()
-	// Two root Describe(Ordered) blocks share one nested cluster (Common Scheduler creates it; Sds reuses in BeforeAll; AfterSuite cleanup); fail fast in CI.
 	if os.Getenv("CI") != "" {
 		suiteConfig.FailFast = true
 	}
 	reporterConfig.Verbose = true
 	reporterConfig.ShowNodeEvents = false
-	RunSpecs(t, "sds-node-configurator e2e", suiteConfig, reporterConfig)
+	RunSpecs(t, "Sds Node Configurator Suite", suiteConfig, reporterConfig)
 }
